+++ kern/include/video.h
@@ -13,8 +13,8 @@ position draw_char_pt(position pos, struct limine_framebuffer *fb, char c);
 
 void print_char(struct flanterm_context *ft_ctx, char c);
 void print_str(struct flanterm_context *ft_ctx, char *str);
-void print_int(struct flanterm_context *ft_ctx, int i);
-void print_hex(struct flanterm_context *ft_ctx, int num);
+void print_int(struct flanterm_context *ft_ctx, long i);
+void print_hex(struct flanterm_context *ft_ctx, long num);
 
 #define ANSI_COLOR_RED     "\x1b[31m"
 #define ANSI_COLOR_GREEN   "\x1b[32m"
